requirements.txt
playwright==1.41.0
PyPDF2==3.0.1
Pillow==10.2.0
requests==2.31.0
beautifulsoup4==4.12.3
openai==1.30.1
python-dotenv==1.0.0
pdf2image==1.16.3
